+++ __tests__/RouterView.spec.ts
@@ -39,6 +39,7 @@ const routes = createRoutes({
       {
         components: { default: components.Home },
         instances: {},
+        enterCallbacks: [],
         path: '/',
         props,
       },
@@ -56,6 +57,7 @@ const routes = createRoutes({
       {
         components: { default: components.Foo },
         instances: {},
+        enterCallbacks: [],
         path: '/foo',
         props,
       },
@@ -73,12 +75,14 @@ const routes = createRoutes({
       {
         components: { default: components.Nested },
         instances: {},
+        enterCallbacks: [],
         path: '/',
         props,
       },
       {
         components: { default: components.Foo },
         instances: {},
+        enterCallbacks: [],
         path: 'a',
         props,
       },
@@ -96,18 +100,21 @@ const routes = createRoutes({
       {
         components: { default: components.Nested },
         instances: {},
+        enterCallbacks: [],
         path: '/',
         props,
       },
       {
         components: { default: components.Nested },
         instances: {},
+        enterCallbacks: [],
         path: 'a',
         props,
       },
       {
         components: { default: components.Foo },
         instances: {},
+        enterCallbacks: [],
         path: 'b',
         props,
       },
@@ -122,7 +129,13 @@ const routes = createRoutes({
     hash: '',
     meta: {},
     matched: [
-      { components: { foo: components.Foo }, instances: {}, path: '/', props },
+      {
+        components: { foo: components.Foo },
+        instances: {},
+        enterCallbacks: [],
+        path: '/',
+        props,
+      },
     ],
   },
   withParams: {
@@ -138,6 +151,7 @@ const routes = createRoutes({
         components: { default: components.User },
 
         instances: {},
+        enterCallbacks: [],
         path: '/users/:id',
         props: { default: true },
       },
@@ -156,6 +170,7 @@ const routes = createRoutes({
         components: { default: components.WithProps },
 
         instances: {},
+        enterCallbacks: [],
         path: '/props/:id',
         props: { default: { id: 'foo', other: 'fixed' } },
       },
@@ -175,6 +190,7 @@ const routes = createRoutes({
         components: { default: components.WithProps },
 
         instances: {},
+        enterCallbacks: [],
         path: '/props/:id',
         props: {
           default: (to: RouteLocationNormalized) => ({
@@ -247,6 +263,7 @@ describe('RouterView', () => {
         {
           components: { default: components.User },
           instances: {},
+          enterCallbacks: [],
           path: '/users/:id',
           props,
         },
@@ -291,5 +308,119 @@ describe('RouterView', () => {
   it('can pass a function as props', async () => {
     const { wrapper } = await factory(routes.withFnProps)
     expect(wrapper.html()).toBe(`<div>id:2;other:page</div>`)
+  })
+
+  describe('warnings', () => {
+    it('does not warn RouterView is wrapped', async () => {
+      const route = createMockedRoute(routes.root)
+      const wrapper = await mount(
+        {
+          template: `
+        <div>
+          <router-view/>
+        </div>
+        `,
+        },
+        {
+          propsData: {},
+          provide: route.provides,
+          components: { RouterView },
+        }
+      )
+      expect(wrapper.html()).toBe(`<div><div>Home</div></div>`)
+      expect('can no longer be used directly inside').not.toHaveBeenWarned()
+    })
+    it('warns if KeepAlive wraps a RouterView', async () => {
+      const route = createMockedRoute(routes.root)
+      const wrapper = await mount(
+        {
+          template: `
+        <keep-alive>
+          <router-view/>
+        </keep-alive>
+        `,
+        },
+        {
+          propsData: {},
+          provide: route.provides,
+          components: { RouterView },
+        }
+      )
+      expect(wrapper.html()).toBe(`<div>Home</div>`)
+      expect('can no longer be used directly inside').toHaveBeenWarned()
+    })
+
+    it('warns if KeepAlive and Transition wrap a RouterView', async () => {
+      const route = createMockedRoute(routes.root)
+      const wrapper = await mount(
+        {
+          template: `
+        <transition>
+          <keep-alive>
+            <router-view/>
+          </keep-alive>
+        </transition>
+        `,
+        },
+        {
+          propsData: {},
+          provide: route.provides,
+          components: { RouterView },
+        }
+      )
+      expect(wrapper.html()).toBe(`<div>Home</div>`)
+      expect('can no longer be used directly inside').toHaveBeenWarned()
+    })
+
+    it('warns if Transition wraps a RouterView', async () => {
+      const route = createMockedRoute(routes.root)
+      const wrapper = await mount(
+        {
+          template: `
+        <transition>
+          <router-view/>
+        </transition>
+        `,
+        },
+        {
+          propsData: {},
+          provide: route.provides,
+          components: { RouterView },
+        }
+      )
+      expect(wrapper.html()).toBe(`<div>Home</div>`)
+      expect('can no longer be used directly inside').toHaveBeenWarned()
+    })
+  })
+
+  describe('KeepAlive', () => {
+    async function factory(
+      initialRoute: RouteLocationNormalizedLoose,
+      propsData: any = {}
+    ) {
+      const route = createMockedRoute(initialRoute)
+      const wrapper = await mount(RouterView, {
+        propsData,
+        provide: route.provides,
+        components: { RouterView },
+        slots: {
+          default: `
+          <keep-alive>
+            <component :is="Component"/>
+          </keep-alive>
+        `,
+        },
+      })
+
+      return { route, wrapper }
+    }
+
+    // TODO: maybe migrating to VTU 2 to handle this properly
+    it.skip('works', async () => {
+      const { route, wrapper } = await factory(routes.root)
+      expect(wrapper.html()).toMatchInlineSnapshot(`"<div>Home</div>"`)
+      await route.set(routes.foo)
+      expect(wrapper.html()).toMatchInlineSnapshot(`"<div>Foo</div>"`)
+    })
   })
 })