+++ src/matcher/index.ts
@@ -16,9 +16,7 @@ import {
   _PathParserOptions,
 } from './pathParserRanker'
 import { warn } from '../warning'
-import { assign } from '../utils'
-
-let noop = () => {}
+import { assign, noop } from '../utils'
 
 interface RouterMatcher {
   addRoute: (record: RouteRecordRaw, parent?: RouteRecordMatcher) => () => void
@@ -323,6 +321,7 @@ export function normalizeRouteRecord(
     instances: {},
     leaveGuards: [],
     updateGuards: [],
+    enterCallbacks: [],
     components:
       'components' in record
         ? record.components || {}